Return typed expressions from expr_error and expr_any
Summary:
Currently, we cannot distinguish between expressions typed as Terr and Tany in the Typed AST, since we represent both with implicitly typed expressions (i.e., the `ty option` in their `ExprAnnotation.t` is `None`). We will likely want to be able to do this when typechecking the TAST.

Additionally, having a type on these nodes helps us make the guarantee that when the `ty option` on an expression is `None`, it is because that expression does not have a meaningful type. Currently, aside from error and any expressions, these are the expressions we do not annotate with a type:

- The list of assignments in a `using` clause
- The member ID in an Obj_get expression
- The Class_const `parent::__construct` in a parent construct call
- `AFvalue`s occurring in key position (which [[ diffusion/FBS/browse/master/fbcode/hphp/hack/src/typing/typing.ml;52f5d74cdb09984b3f772b52e9f8a58581811f65$3130-3132 | shouldn't happen ]])
- The IDs of special functions (`idx`, `array_map`, `array_filter`, etc.)
- The IDs of hh pseudofunctions (`hh_show`, `hh_show_env`, `hh_log_level`)
- The IDs of PHP/Hack pseudofunctions (`unset`, `empty`, `freeze`, etc.)

We might choose to annotate some of these with a type in the future (particularly the special functions, where their type is specialized by the way they are invoked), but others, like the ID in an Obj_get, never have a meaningful type.

Reviewed By: andrewjkennedy

Differential Revision: D7111503

fbshipit-source-id: 13d9fe8031e870c9be69c2e126240e6e24d5081e