Add ExprAnnotation to fun_params
Summary:
Currently, fun_params are not annotated with a type in the Typed AST. Consumers who would like to know the type of a function parameter need to localize the hint, which doesn't give them instantiated generics, and also doesn't help when no hint is present.

This change annotates the fun_param type with the AAST's ExprAnnotation. This is a little redundant, since a fun_param already has a param_pos, and the NAST and TAST ExprAnnotation types contain a pos. It's more straightforward than adding a new ParamAnnotation type to the AAST, though. Perhaps we could move the Pos.t out of the NAST and TAST's ExprAnnotation.t later.

Reviewed By: pittsw

Differential Revision: D7102916

fbshipit-source-id: 73d0697126f208f893f97274304821c17a424659