Ignore naming errors during decl
Summary:
We used to run local naming only once, as a separate phase before decl, and store named ASTs in shared memory to be re-used by decl and typing.

A long time ago, as a memory saving feature, we stopped doing it, and instead run naming before both decl and typing. This causes all naming errors to be reported twice. It's mostly invisible to uses, because we deduplicate errors before sending them.

It does create more surface area for making mistakes (we need to ensure that we correctly update both versions), so let's instead ignore one of them. I choose to ignore the decl ones, because in D7094942 I will start storing them in heap, so why not to reduce their size. I also think it's cleaner to return as many errors as possible from typing, and treat decl as optimization implementation detail.

This changes reported errors from:

  naming -> decl -> naming -> typing

to

  decl -> naming -> typing

`hh_single_type_check` has this annoying property of displaying only the first encountered error, so if there was both naming and decl error, it will now change the test output. This makes it very confusing to asses whether change is safe, so I choose to instead HH_FIXME cascading errors to maintain the output.

As a side note: does the order and all this effort in maintaining it matters? It's not clear...

- all user visible errors are returned through `Errors.get_sorted_error_list`, so the only ordering that matters is of errors that are exactly in same position
- there is `Errors.try_with_result`, which tries something, and if it fails, calls a function with "last error". If we change it, it might completely change some error messages. But this happens only within a single phase, so switching order of naming and decl as this diff will not matter.

I'm not at all convinced if any order matters, but it's easier to keep maintaining.it than prove that it doesn't.

(Note: this ignores all push blocking failures!)

Differential Revision: D7094909

fbshipit-source-id: aa58fa711c69199ce462af97980ab1a2e45d6cfc