Fix Errors.merge to *really* behave like List.rev_append
Summary:
While a list has only one neutral element - empty list, a map of lists can have both an entry with empty list, or no entry at all. Usage of Option.merge was not the right choice if we wanted the merging of maps to behave similarly to merging of lists (see new test case that would fail before - the errors would not be reversed).

As mentioned in {D7094909} - no idea if this whole ordering matters, but test outputs change if the ordering changes, so it's easier to maintain it.

(Note: this ignores all push blocking failures!)

Reviewed By: pittsw

Differential Revision: D7094911

fbshipit-source-id: 344e9740e23b5c8d5257a6e89df5f76c6e510914