Errors.merge_into_current
Summary:
We can get two sets of errors and Errors.merge them - but that requires getting all the errors at the callsite, while it's easier and cleaner to get them inside the callee (see application in D7094942). This is a helper function to make it possible.

(Note: this ignores all push blocking failures!)

Reviewed By: pittsw

Differential Revision: D7094913

fbshipit-source-id: 0b47f81a81f286d9bc995330ff4bbb5f322ef5b1