Do not return errors from decl - store them in heap and return from typing
Summary:
It's possible that typing of file A will return typing errors for A, and possibly (and quiet nondeterministically) decl errors from some of it's dependencies.

To make it more deterministic, I change it so that the decl errors in B are returned only when typing B. If they happen to be computed when typing A, they will not be returned, but stored along with B declaration.

This assumes that our dependency tracking system will always tell us "you need to recheck B" whenever it can have new decl errors (and will not tell us to recheck A instead and depend on this recursive thing). I believe it to be the case.

Class element errors remain stored with their class (since at the moment it's impossible to declare only a single method/property - it's always the entire class).

(Note: this ignores all push blocking failures!)

Reviewed By: jamesjwu

Differential Revision: D7094942

fbshipit-source-id: 6519714cbe9dc9b698f48bcfc7debfaa379e3c67