Use hash_int64 in the Fixup TreadHashMap
Summary:
We originally argued that since x64 instruction lengths are arbitrary,
the identity hash would be fine. And thats probably true(ish) for
x64. For PPC and Arm, instructions are all 4 bytes, so when the table
is 75% full, the average chain length will be 3.

Use hash_int64 instead (which is implemented by the hardware crc32
instruction when available). This will probably be an improvement on
x64, and will definitely be an improvement on Arm and PPC.

Reviewed By: ricklavoie

Differential Revision: D7115026

fbshipit-source-id: 2d3b3641386076f7af0dbf9a1ef0d7d2e7bf8e3b