Don't filter out hover responses which would display an underscore
Summary:
Currently, ClientLsp filters out hover responses of `None` and hover responses which would display `"_"` for the type under the cursor. These responses are slightly different, however--`None` indicates that there is no type at all (e.g., the cursor is over a comment), but `"_"` indicates that the type of the hovered expression is either `Tany` or `Terr`.

I think it's better to display an indicator that the expression is untyped than to not show the datatip at all (and leave the user wondering whether the datatip failed to show because the expression is untyped or because Nuclide or Hack has a bug).

Reviewed By: pittsw

Differential Revision: D7140239

fbshipit-source-id: a2b4d86ff1b752f6b587aa20c90c0ebb7fc0f4d8