Escape XHP content
Summary:
Text of XHP body and attributes should be processed to decode HTML entities however escape sequences in it are not handled.

Note: In existing code during lowering we conflate several cases that should be processed differently, i.e.
```
<a attr="\t &amp;"/>
```
and
```
<a attr={"\t &amp;"}>
```
is both turned into
```
Xml { attributes = String("\t") }
```
In order to deal with it correctly decoding of HTML entities is moved to lowering

Reviewed By: oulgen

Differential Revision: D7149685

fbshipit-source-id: 44274c0cc03e0c25b35b6322770d93162b417170