Check realpath before returning cached unit in non-repo mode
Summary:
If we have a cached unit at a path containing a symlink and the link
changes to a new file that's a hardlink of the old file, we will erroneously
hit in the cache. Add a realpath check to the lookup path to fix this.

Everything about the cached unit is OK except for its filepath, so if this
causes any problems we can look into smarter solutions that don't duplicate the
whole unit.

Reviewed By: markw65

Differential Revision: D7148499

fbshipit-source-id: cd2b6fbd916bfa09f87c48dd0f48c9acb432e0c5