getrusage
Summary:
Recently I was debugging some worker perf issue, and I wanted
to know how often the workers were being voluntarily or involuntarily
context switched by the kernel. This information is available using the
[getrusage](http://man7.org/linux/man-pages/man2/getrusage.2.html)
system call.

I thought this might be worth checking in, in case anyone else needs
this in the future.

Reviewed By: gregory-nisbet

Differential Revision: D7075560

fbshipit-source-id: 4306ee8bae05284c219c6cb38ee9ec29004cda10