Safe instanceof should elide nullable
Summary:
Currently it's the case that `instanceof` tests on locals that have a nullable type do not glean
concrete information about type parameters. For example, if `$x` has type `?Base<T>`, and we do `$x instanceof Derived`, then the type of `$x` is refined to `Derived<T#1>` with no information about `T#1`. In the case that `Base` is invariant, it should be possible to determine `T=T#1`. To fix this, we add a new rule to the decomposition of a subtype assumption: if we know that `C<ts> <: ?D<ts'>` for classish types `C` and `D`, then it must be the case that `C<ts> <: D<ts'>`.

Reviewed By: vladima

Differential Revision: D7153433

fbshipit-source-id: 898d9e6443c5f1097c21acef7a00d6d6820a5f47