Remove AttrHot, add AttrDynamicallyCallable
Summary:
The "hot" attribute on functions does not actually need to be an Attr. Its only
manipulated from within the runtime, so it does not need to be stored in the
repo or in FuncEmitter. The only real need for it to be in an Attr was support
for EvalRandomHotFuncs and FunctionSections, neither of which are supported by
HackC nor seem to be needed any longer. Instead use a bool in Func to store
hotness.

With the bit freed up by removing AttrHot, add AttrDynamicallyCallable and
remove the bool indicating dynamic callability from Func. Not only does this
make the logic closer to how we deal with other attributes, but it opens the way
for JIT generated code to query dynamically callability quickly.

Reviewed By: markw65

Differential Revision: D7142592

fbshipit-source-id: cfc0e0e56688f16362976fad1c7a3c6dca4a8779