Improve error messages for dynamic types
Summary:
This makes error messages for dynamic types cleaner, adding two new reasons: accessing a property of a dynamic type yields another dynamic, and calling a dynamic type like a function yields another dynamic type.

Currently, assigning a dynamic type to a lvar does not change the reason information, so it will still point to the source. This is also true of regular assignments as well, but leads to interesting behavior like:

```
function dynamic() : dynamic {...}
function test() : int {
  if(...) {
    $x = dynamic();
  } else {
   $x = 5;
 }
  return $x;
}

```

This will raise an error pointing at line 1, rather than line 4.

This is true if you replace dynamic with a string, too, though, so, this is more a question of which gives more useful info.

Reviewed By: dlreeves

Differential Revision: D7091619

fbshipit-source-id: 43ab279f53fd00c4821ce36339c37a3399a403a1