Remove incremental_errors flag
Summary:
This option was there just in case something went terribly wrong and I wanted
to back out of this feature without having to back out the entire release. It
has been enabled for everyone for almost two months[1] and it's unlikely we'll
want to suddenly go back on this.

Reviewed By: alexchow

Differential Revision: D7165454

fbshipit-source-id: 732ec9df5a5f5e9ce0df19cd441f9ef63ad0cab9