Don't double free ephemeral stubs
Summary:
There was some code to attempt to detect the case where a stub was
freed multiple times because more than one thread was running through
it, and not push it back onto the free-list more than once.

The code was broken, because if it ever happened, the stub could be
pushed, popped, and then pushed again, leaving a live stub that was
also on the freelist (and either corrupting the freelist when the stub
got jitted, or corrupting the stub when it was pushed onto the
freelist, depening on which happened first).

After inspecting the code, it looks like its already very careful not
to smash the same stub twice, so just always_assert if it happens.

Reviewed By: ottoni

Differential Revision: D7180534

fbshipit-source-id: 66afadc688c293df6952352bb33f5a9ed046047d