Properly handle closing stdin/stdout/stderr for cli-server
Summary: Trying to `fclose(STDOUT)` in cli-server mode was properly closing the `FILE*` resource but allowing the server to perform a double free when the `RemoteFile` wrapper itself called `fclose`. To mitigate this, transfer ownership of these files over to `plain-file.cpp` as they are only used for `BuiltinFile` resources. This way we can ensure that they are only ever closed once.

Reviewed By: swtaarrs

Differential Revision: D7176703

fbshipit-source-id: 738f262cc386a9777cad7872159145758068d6de