Codemod imports from libfb to use full path /havoc-/insights
Summary: This is only fixing imports of the format "from libfb.X import Y" -- this is a little easier because there are no further references to libfb.X that need to be fixed in the file. There are many diffs to follow doing the same thing + another codemod pass coming.

Reviewed By: eugeneoden

Differential Revision: D7177865

fbshipit-source-id: c9252980c49ab63889d79398467fbc07f831a660