HHVM Debugger: Adjust breakpoint calibration algorithm
Summary:
This adjusts the breakpoint calibration algorithm to be more intuitive about the source location it selects to insert a breakpoint at in certain cases.

Before, our rule was to pick the source location whose startLine >= breakpoint request line that has an end line nearest to the breakpoint request line.  This is sub-optimal for certain types of code.

For example:

  1.  $z = foo(
  2.      $a,
  3.      $b,
  4.      $c
  5.    );

If a bp is requested on line 1, the previous algorithm would put it on line 2, which is the evaluation of the first argument.  The user actually requested a bp on $z = <return value of foo()>, which should really end up on line 5 in the UX. The matching source line is the source location that begins on line 1 and ends on line 5, not the source location that begins and ends on line 2.

Note that in practice this doesn't make a huge difference when debugging, because either way we stop before evaluating the first parameter of the function call.

If the source looked like this instead, we'd pick line 1:

  1.  $z = foo($a,
  2.      $b,
  3.      $c
  4.    );

Reviewed By: velocityboy

Differential Revision: D7187661

fbshipit-source-id: 96abc16109a4fbd587c3b07d376167a9d412ce9e