Add runtime option to enable/disable XED
Added a runtime option to enable/disable XED when using the XED
library: Eval.UseXedAssembler. Previously this was a build-time
option which chose one of two versions of the X64Assembler class
to be used (either the one based on XED or the previous
implementation).
The following changes have been made in order to make
switching between the two implementations a runtime option:
- The XED-based X64Assembler has been renamed to XedAssembler
- Both X64Assembler (which contains the previous implementation) and
  XedAssembler now share a base class: X64AssemblerBase.
- Their public interface is now identical.
- X64AssemblerBase is an abstract class which defines the interface
  shared by its two subclasses. All common code from the two subclasses
  has been moved to this base class.
- struct Vgen from vasm-x64.cpp is now a template and will be used
  with either XedAssembler or X64Assembler, depending on the runtime
  option.
- In all other places where X64Assembler was used and the use of
  templates is not an option, the NEW_X64_ASM macro is now used to
  instantiate one of the two subclasses, depending on the runtime
  option.
XedAssembler incorrectly used CodeBlock::frontier() instead of
CodeBlock::dest() as a destination buffer for XED writes. This
is now fixed.