Speedup ForbidDynamicCall implementation
Summary:
The existing dynamic call checking infrastructure has high overhead when turned
on, mainly because it inserts a dynamic call check in the prologue of every
function. Improve this by moving the dynamic call check to the caller (instead
of the callee). When making a function call which is known to be dynamic, check
if the target function is dynamically callable, and if not, emit the warning.

The dynamic call checks for frame accessing functions and for builtins still
remain in the prologue. Its more optimal to do those there as they benefit more
from knowing the func statically. Therefore we now have "caller dynamic call
checks" and "callee dynamic call checks".

Reviewed By: paulbiss

Differential Revision: D7170451

fbshipit-source-id: d8378ae4e82876fe04bb73f35e4a91dd9fd6a011