Fix crash merging unreachable state while inlining
Summary:
The unreachable code generated after the ThrowInvalidArrayKey
instruction had the wrong stack offset. Fix it by committing the stack effects
of the AKExists before throwing.

I have a diff in progress to actually fix the underlying issue with unreachable
code, but it's a much bigger, more invasive change. This is a fix just for the
crash in the test case I added.

Reviewed By: markw65

Differential Revision: D7227352

fbshipit-source-id: a7d7fe4045370c4e7eb9aca51c2e8e9ae6629ff8