HHVM Debugger: Include request URI in threads response
Summary: If available, include the request URI in the thread name in a response to a ThreadsCommand from the debugger client. Since the server usually has tons of requests in flight, even for a single browser client, this makes it easier for the user to see which request is which.

Reviewed By: mostafaeweda

Differential Revision: D7217836

fbshipit-source-id: dcf8b0ffb0699e6f200fc4f4bb4cdf3bff8f39a8