Type invocation of unresolved union containing only one type as if it were unwrapped
Summary:
Currently, the arguments to an invocation of an unresolved union are discarded in the TAST. This is because we produce the TAST nodes for the arguments as an output of the `call` function, since their type can be influenced by context.

To deal with this in the case of an unresolved union of only one type, this change has us typecheck that call as if the function type was not wrapped in a `Tunresolved`.

Reviewed By: andrewjkennedy

Differential Revision: D7191958

fbshipit-source-id: 92f84b3710c65588956ff5157db8e350853c3c44