Fix ref case in Variant detach helpers for objects and resources.
Summary: Variant::detachResourceData and Variant::detachObjectData got the refcounting wrong in the case where the Variant contained a ref. I can't find any code that can trigger this, because we tend to pass Variants around by value in the code that casts them to req::ptrs, so they get ref demoted and don't hit the bad code path. Fix it anyway in case there is a code path I couldn't find that can trigger this.

Reviewed By: markw65

Differential Revision: D7217569

fbshipit-source-id: 6378f1b8931f6b18c443e4cb0ea9e3e088b1d00a