Put TreadHashMap on huge pages, and allocate lazily
Summary:
SrcDB lookup seems to be hot, espeically during warmup. This tries to put it on huge pages to reduce tlb misses.

Allocate on huge pages lazily, so that we don't have to create the huge page arena at the beginning (need to defer creation of static strings, etc. as well later).

Reviewed By: markw65

Differential Revision: D7154233

fbshipit-source-id: 902e1af752c96cfc15d0ec5775d285c45d9b1744