Use type history of terminal try/catch blocks when typechecking finally block
Summary:
When a `finally` block is present, it is not correct to omit terminal `try`/`catch` blocks when intersecting their environments, since any statement in the terminal blocks could throw (thus entering the `finally` block). The `finally` block should intersect the history of local types throughout all `try`/`catch` blocks, including terminal ones.

However, for typechecking statements following the `finally` block, we would like not to include the type history of terminal blocks in the intersected types of locals, since it is not possible for locals to have those types after the end of the finally block.

To deal with this, this change typechecks the finally block twice. First, with the intersected local history of all try/catch blocks (including terminal ones). The typed AST for the finally block which we produce in this pass is used in the TAST for the `Try` node. We then typecheck the finally block a second time, but exclude the type history of terminal blocks from the intersected types of locals.

Because I found `intersect_list` to be a misleading name, this change renames it to `intersect_nonterminal_branches`. A new function (private to Typing_lenv) named `intersect_list` is added which does a straightforward intersection of a list of local environments.

Reviewed By: jamesjwu

Differential Revision: D7093698

fbshipit-source-id: a39ef8f34f8ed0ee58211cd6c54aaad037ddadc0