not assuming the number of jemalloc arenas is small
Summary:
Previously, the implementation of ManagedArena assumes the number of jemalloc arena is very small (it is 1 by default), so an array indexed by arena id is used to obtain a pointer to the arena given an id.
However, it crashes when MALLOC_CONF sets it large and when ManagedArena is used.
This fixes it by using an array of pair<id, pointer> to store the mapping between arena id and pointers to the arena.

Reviewed By: interwq

Differential Revision: D7206066

fbshipit-source-id: abbfcb573f2bc2581f6b7a8b6927c208a51fb17f