An initial implementation of Dynamic View
Summary:
This diff adds a new mode to hh_single_typecheck, "--dynamic-view", which will replace all instances of Tany in the typechecker with Tdynamic(thus introducing strictly more errors).

The final goal of this is the ability to run most IDE commands(type-at-pos, get errors being the two main ones) with an extra flag "--dynamic-view", replacing Tany with dynamic.

An important invariant here is that any file which typechecks with --dynamic-view on should typecheck with dynamic-view off, which makes dynamic view strictly less permissive than regular Hack.

Implementation details:

The main change is in typing_phase, where localizing a Tany or Terr results in a Tdynamic when dynamic view is on. Other than that, I replace any place where we explicitly construct a Tany or Terr in typing.ml with a utility function which returns Tdynamic when dynamic_view is turned on.

I need to do this for other files too, like Typing_* files, but I'll probably add those separately to reduce the size of these diffs.

Reviewed By: oulgen

Differential Revision: D7247056

fbshipit-source-id: 1055dc742190148dc8ba76b199dd1844b7395feb