Replace various usages of Reason.Rnone with Reason.Rwitness, add a few more tanys
Summary:
- We used to use Reason.Rnone very often with Tany, since there's generally no way for Tany to raise an error and use that information. But with dynamic view, we need the position information to give better types. I've looked through all the Reason.Rnone in typing.ml and moved the ones I believe are correct to Reason.Rwitness.

There are a few places where Rnone is used intentionally, say to represent the lack of any "self" or "this" type. I avoided those for now, but in my opinion, those too should move from trying to pattern match on Reason information to match on the type itself, or better yet, an option type. Reason information should not be used for type inference.

- Add TUtils.tany to a few more places that need them.
- Fix a bug with bind_params

Reviewed By: andrewjkennedy

Differential Revision: D7260975

fbshipit-source-id: a4519ef46ddd4b73b26b705305bc851974e58307