FFP implementation of language-and-mode parsing
Summary:
For some tooling, we only want to know the language (php/hh) and mode (partial,
decl, strict) of a file. For others, the parser configuration depends on it, so
it would be nice to have a way to quickly grab language and mode before
starting the "proper" parse.

This diff adds a quicker way to extract language and mode from a file by means
of the full fidelity parser.

Reviewed By: pittsw

Differential Revision: D7155247

fbshipit-source-id: eddefcf60f83de99fd4d32a64eb492bb72caa810