Add smart constructors for decl mode
Summary:
We want to use smart constructors in the decl mode. Idea is to drop everything in the function bodies and extract only information whether it's a generator(contains yield) or not.

These smart constructors are doing exactly this. If the function is a generator then a whole body would be replaced with `yield` token otherwise it will be `missing`.

Here is an example of a tree parsed with these smart constructors.

```
...
> > > > simple_type_specifier
> > > > missing
> > > function_declaration_header
> > > > {
> > > > missing
> > > > }
> > > compound_statement
> > function_declaration
> > > end_of_file
> > end_of_file
> list
script
```

(Note: this ignores all push blocking failures!)

Differential Revision: D7278446

fbshipit-source-id: 721261476cc2c14ea7c931ebd4354f66e0e93148