Make parameters with no default argument default to Tany
Summary:
When a parameter has no type hint, we create an empty type variable so that we can subtype it with its default parameter to get a better type for it. This actually is too permissive(and also not really correct since nothing guarantees $x is going to be followed, but that's a story for another day):
```
function f($x = 5) : int {
  return $x; // $x is an int here, but should probably be num
}

function breakIt() : void {
 takes_int(f(5.4));
}
```

Whether or not this behavior is intended or correct(which should probably be addressed in a different diff), when a parameter has no default argument OR hint, this empty type variable has no use, and is equivalent to Tany.

This just makes it so that we enforce the parameter as dynamic in dynamic_view mode, when no default argument or hint is given.

Differential Revision: D7280450

fbshipit-source-id: c16591995f6e41ace4c7c92a2ecd995593c75293