Ping non-persistent clients before processing their commands
Summary:
It's possible for client to be queued so long that it goes away (by user doing Ctrl+C or some combination of `--timeout` `--retries` that happens to work) before server even starts processing its request. In this diff, I ping clients just before we start doing it to detect this case and save us doing that work.

It used to be that server saying HELLO would effectively serve as such ping, but as explained in D7178015 it sometimes doesn't work that way and it will become more and more likely in the future.

I ping only Rpc clients, because Debug is used only in tests, and BUILD (going away) is the only actually used Stream command.

Reviewed By: gregory-nisbet

Differential Revision: D7207801

fbshipit-source-id: 672cc611fcd2afda1e47dd8f47942748f17c60a5