Fix + add more logging
Summary: Before we were not capturing the case that conn_op.ok() returned false, this causes segfaults later on. On top of this I created a new event to store both the connection op and query op, this enables us to return timing information for both of them on the success path. Note for the failure path we only return the info for 1 or the other.

Reviewed By: jrahman

Differential Revision: D7253062

fbshipit-source-id: 1948fc08a4c0ea1d38775bcabd16893f8dce5d9d