Implement dynamic view for non LSP type-at-pos
Summary:
This diff adds support for the following:

```
hh_client --type-at-pos file:line:col --dynamic-view
```

Which runs type_at_pos assuming dynamic-view is on.

I'll be adding more and more features as I go on, but this is the easiest one to do for now. Later, I'd like to support opening an ide file in "dynamic view" mode, which enables dynamic_view errors and type-at-pos for the file.

Differential Revision: D7280451

fbshipit-source-id: 46cac8c08e10198e15d62328bded59d328d8bae8