Check for Worker oom in coalesced failures
Summary:
Things were set up in D7210619 to coalesce failures from workers.

Quick review of the problem, our data was wrong and we under-report OOMs
because OOMs cascade into other problems via corruption of SharedMem. So
by chance we encounter crashed workers before pinging the result of the
OOM'd worker, and end up logging a crash instead.

This fixes that.

Reviewed By: dabek

Differential Revision: D7259486

fbshipit-source-id: 5815c227e874608a3430e4155a2784f9bd69bcaf