Remove hh_client ide .
Summary:
It was completely superseded by `hh_client lsp .`

Parts of it (`nuclide_rpc_message_printer.ml`) are still used to serve `--json` versions of the command-line queries. I removed everything around it and left minimal set of functions necessary for it. They could probably be refactored and moved around to look nicer, but meh... it's still better than with a bunch of unused cruft around.

Reviewed By: alexchow

Differential Revision: D7208841

fbshipit-source-id: 6274151765d3b3db3ada3dc5b7034439e088f8e4