Watchman should raise exception when the subscription is canceled
Summary:
When Watchman cancels a subscription, it sends a "canceled: true" in the JSON PDU.

We have been ignoring this.

Watchman also explicitly does not close the socket connection, because there might
be multiple subscriptions on a socket (even when there isn't). So a EOF will never show up on the socket.

Since we ignore the Canceled PDU, we think the subscription is still alive when it
isn't, and we never see any further events on it. This can lead to
watchman_event_watcher getting wedged.

Reviewed By: wez

Differential Revision: D7298939

fbshipit-source-id: a664ffb201f4431be32cfc90a6e758ff096a7ca4