Flow lsp can report telemetry/connectionStatus
Summary:
Most integration tests want to wait until the flow lsp adapter has connected to the flow server, before they start issuing requests.

Previously that information was only indirectly available via human-readable progress reports and dialogs.

With this diff, the test can indicate that it supports the telemetry/connectionStatus capability, and the LSP adapter will send whenever its connection status changes. (the initial default is false).

Note: I had the choice of whether to opt into this with a command-line argument or an LSP capability. I chose the capability because that's more uniform with the rest of LSP.

Reviewed By: gabelevi

Differential Revision: D7275740

fbshipit-source-id: 82dcc644d4aa56786d82f112fcee968c070dba5b