Make Hack LSP return ranges
Summary:
Back when Hack LSP used TYPE_AT_POS to return hover data, we couldn't get positional data out to return to the client. Because of this, Nuclide would fall back to using only the single hovered-over glyph as the location, which would a) make it nearly impossible to move the mouse over the type hint box, and b) mean that we didn't get useful highlighting over the entire identifier.

Now that we use the dedicated IDE_HOVER command to get hover details, we can change that to also give us positional data.

Differential Revision: D7312548

fbshipit-source-id: 29a24182a4de4187af9a01ba5bd73f05e004d129