Set g_context's transport earlier during request init
Summary:
Some upcoming extensions want to inspect it at request-init time.

Nothing currently observes the transport between where it's set now and
where it was being set before---and the only exceptional condition that
arises now which didn't before is that the tranport will be set if one
of the obStart() calls in executePHPRequest() throws.  However, in that
event, no PSP functions will have been registered, so nothing in the
catch will observe the set transport either.

Reviewed By: markw65

Differential Revision: D7318556

fbshipit-source-id: f14f60b3f5e10f211211a6f9f311633474a8b298