Rewrite facts parser to use smart constructors
Summary:
Initial stab of switching facts extraction to use smart constructors.
The overall approach still stays the same: we parse source text into structured form and then analyze it to extract facts about the source code. A difference is that instead of using full fidelity tree we create compact structure that effectively stores only declarations that contribute to facts. As a result tree structure we get is very shallow and thus can be processed faster.

Differential Revision: D7305294

fbshipit-source-id: 0af298139dc2329250f727a18aeb865d12a1c234