HHVM Debugger: Rotate debugger log files based on max file size
Summary: Using age is not reliable on all file systems, switching to cap log files at 0.5MB and keep the most recent 5, rather than trying to keep the most recent 5 days worth of logs. NOTE: Logging for the debugger is disabled completely by default, and is enabled either in HHVM config, or opted-in by the debugger client when it connects.

Reviewed By: velocityboy

Differential Revision: D7258141

fbshipit-source-id: 730c576bb0e5e1b8614f6b045cd46e31ef861132