Handle classname<this> typehints
Summary: This uses the definition of `this` that HHVM uses, i.e. that it's exactly the LSB classname (not a subtype), so we do an equality check instead of `is_a()`. Relatedly, it may be useful for us to add a `static` typehint to Hack to represent what `this` used to be known as.

Reviewed By: mofarrell

Differential Revision: D7132884

fbshipit-source-id: 67996d12341f842bd7c48dcb43068194ba14b13d