Add autocomplete integration tests
Summary:
I want to make a fix for some aspects of autocomplete. This diff doesn't change anything; it just adds unit-tests that embody the current status quo.

I've added a new feature to test_py, to cope with tests that emit a path to the builtins temp directory. I've also removed the maxDiff length, in the hope that sandcastle will now be able to show full test logs in all cases.

Reviewed By: pittsw

Differential Revision: D7332189

fbshipit-source-id: aa338ffda4da94f55bc71d6bb23a901ce4a87846