Hack no autocomplete after single colon
Summary:
Prior to this diff,
```
MyEnum:|  // this showed autocomplete for globals
case MyEnum::Item:| // this showed autocomplete for globals
```

These were wrong. They shouldn't have shown any autocomplete items. And indeed, showing autocomplete there was interfering with normal typing of case statements.

But we don't want to compromise some other autocomplete experiences after colon...
```
MyEnum::|  // shows members
$xhp->:|   // shows properties
new :|     // shows xhp classnames
```

And there are some autocomplete after colon that are already broken, which we won't fix in the current diff
```
<nt:|   // shows nothing
<nt:t|  // shows nothing
$xhp->:nt: // shows nothing
new :nt:  // shows nothing
```

Anyway, this diff fixes up the first two cases. It fixes them with the heuristic that if we're asking to provide autocomplete after a single colon, then we should only do so if the colon was already part of a prefix. In all other cases, we shouldn't.

It'd be better to do this via FFP autocomplete. I'm doing it now so that (1) we can fix it now, (2) when we switch over to FFP autocomplete, we'll have regression tests in place.

Reviewed By: pittsw, arxanas

Differential Revision: D7332305

fbshipit-source-id: 071b68864274b4857eabb3de4acce47ffd43d6ba