Functorize full_fidelity_ast
Summary:
Currently, the lowering process done in `Full_fidelity_ast` will convert the `full_fidelity_positioned_syntax_tree` to a `full_fidelity_editable_positioned_syntax_tree`. This conversion is an expensive pass and currently is only done to lower coroutines. Thus, it would be beneficial to only perform the conversion when we need to lower coroutines. However, the lower functionality assumes a `full_fidelity_editable_positioned_syntax_tree`. Thus, we need to functorize the lowerer so that it will take either syntax and perform lowering.

Having the `Full_fidelity_ast` lowerer take in `Syntax_sig.Syntax_S` proved to be a little too restrictive. Instead of stubbing out those function declarations in `full_fidelity_minimum_syntax` and corresponding `token` and `trivia`, created new signatures of `full_fidelity_positioned_syntax` and `token` which include additional functions that were not available before for the use of the lowerer. These interfaces were then implemented by both the editable positioned and the positioned flavors of Syntax.

Differential Revision: D7221613

fbshipit-source-id: d710ab30576fe521c558a6c8e1581866cb876161