Logging to track Relative_path.Dummy showing up in unexpected places
Summary:
The assertion in errors.ml is (rarely, compared to memory issues, but still) being hit in production. That means that at some point we are thinking that there are definitions in Dummy ("default") path, which is likely very bad for our incremental mode.

Things should have default path only when we are analyzing something "on the side" for the purposes of IDE, and should not leak outside of this computation thanks to ServerIdeUtils.make_then_revert_local_changes.

They do somehow leak out, or are created elsewhere, and in this diff I add checks in few more places where they should not show up to narrow down how does it happen:

- log when there no local changes (i.e we are not inside make_then_revert_local_changes) but we add a default path to naming heap
- log when there are no local changes and we get the default path back from TypeIdHeap (the only heap with cache). If this happens, try bypassing the cache to check if it matches the actual heap and log if it doesn't
- log when default path shows up in "files to recheck" in ServerTypecheck
- log in which phase are the "bad" errors created

Reviewed By: gregory-nisbet

Differential Revision: D7334282

fbshipit-source-id: 285d076073946ce81e83dd018c6f6e4021003cdb