Split safe_instanceof into individual functions.
Summary:
This diff splits up `safe_instanceof` into functions that can be reused by the `is` expression type refinement code in a followup diff.

It also makes one substative change: **The elements of `tparams_with_new_names` are wrapped in options.** This is because in `is` expressions, we don't generate a fresh type for every type param—we only generate them when we see wildcards (`_`). This means that, by the time we get to `safely_refine_class_type`, we only want to simplify constraints on fresh type params that we added, and leave existing types alone.

Reviewed By: andrewjkennedy

Differential Revision: D7360187

fbshipit-source-id: 1708c9f4153ba6798f1eecf94afa89a9cc5e77af