HHVM Debugger: Install breakpoints correctly on new compilation units
Summary: The debugger needs to re-install breakpoints on new compilation units even in the case where the file path for the unit has already been seen by the debugger, and there are no pending unresolved breakpoints for the current request + unit. If the file changes on disk, or is not in the unit cache for any reason, or is compiled by JIT, we can see a new unit generated for a file path we've already installed breakpoints on. Not placing the bps in the new compilation unit causes them to be missed.

Reviewed By: velocityboy

Differential Revision: D7358170

fbshipit-source-id: 41048fae8d2b4f8a7ba1b9ec4fe625bd183513d7